Clean up tests to minimize duplication
