--- src/Simulation/Simulators.Type2.Tests/Circuits/AssertEqual.qs
@@ -1,12 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-namespace Microsoft.Quantum.Simulation.Simulators.Tests.Circuits {
-    
-    function AssertEqual<'T> (expected : 'T, actual : 'T) : Unit {
-        body intrinsic;
-    }
-    
-}
-
-
