--- src/Simulation/Simulators.Type2.Tests/TestProjects/IntrinsicTests/IntrinsicTests.csproj
@@ -1,30 +1,0 @@
-<Project Sdk="Microsoft.Quantum.Sdk/0.12.20082209-beta">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <IsPackable>false</IsPackable>
-    <ExposeReferencesViaTestNames>true</ExposeReferencesViaTestNames>
-    <!-- we will provide our own -->
-    <CsharpGeneration>false</CsharpGeneration>
-    <IncludeQsharpCorePackages>false</IncludeQsharpCorePackages>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\CsharpGeneration\Microsoft.Quantum.CsharpGeneration.fsproj" PrivateAssets="All" IsQscReference="true" />
-    <ProjectReference Include="..\..\..\Simulators.Type2\Microsoft.Quantum.Simulators.Type2.csproj" />
-    </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.4.0" />
-    <PackageReference Include="xunit" Version="2.4.1" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.1" />
-    <DotNetCliToolReference Include="dotnet-xunit" Version="2.3.1" />
-  </ItemGroup>
-
-  <Target Name="BeforeCsharpCompile">
-    <ItemGroup>
-      <Compile Include="$(GeneratedFilesOutputPath)**/*.g.cs" Exclude="@(Compile)" AutoGen="true" />
-    </ItemGroup>
-  </Target>
-
-</Project>
