--- src/Simulation/Simulators.Type2.Tests/TestProjects/Library2/Library2.csproj
@@ -1,21 +1,0 @@
-﻿<Project Sdk="Microsoft.Quantum.Sdk/0.12.20082209-beta">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.1</TargetFramework>
-    <!-- we will provide our own -->
-    <CsharpGeneration>false</CsharpGeneration>
-    <IncludeQsharpCorePackages>false</IncludeQsharpCorePackages>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\CsharpGeneration\Microsoft.Quantum.CsharpGeneration.fsproj" PrivateAssets="All" IsQscReference="true" />
-    <ProjectReference Include="..\..\..\Type2Core\Microsoft.Quantum.Type2.Core.csproj" />
-  </ItemGroup>
-
-  <Target Name="BeforeCsharpCompile">
-    <ItemGroup>
-      <Compile Include="$(GeneratedFilesOutputPath)**/*.g.cs" Exclude="@(Compile)" AutoGen="true" />
-    </ItemGroup>
-  </Target>
-
-</Project>
